python/requirements.txt
pypdfium2
pdftext
torch>=1.8.0
ipython
ultralytics
onnx
onnxruntime-gpu
onnxconverter-common

ocrmac ; sys_platform == "darwin"
